utils: WriteFile function
